chore: factor out language client implementation and clean up a bunch of stuff related to it
